Implemented a method of differentiation
After messing around with Caciuc's integral a bit and trying to
get it in terms of the potential, I figured this was the next
best strategy.

Although now that I think about it, I don't think this will help
much given the fact that the integral appears to me to be
divergent. :/

Oh well, lateral progress ahoy