Added back numbered methods for axes
Added back methods like coord_1() for each axis (though they now go from
0..n-1 rather than 1..n), and the methods that take an axis as a
parameter now end in _n to discourage their use when the axis is known.

This is to help prevent mistakes such as those fixed in the previous
commit, as the numbered methods throw compile time errors if an invalid
axis is used (such as size_3() on a Lattice3; an easy mistake).