Implement StrictParser
A class that reads data sequentially from an input stream in a
fail-fast manner (i.e. an error is thrown immediately when the
parser encounters something not of the format it expects to find)

Its methods were originally part of my LOCPOT file parser (currently
under construction), but it seemed appropriate that it become a
separate class.