Base implementation of Point and some tests
Point is being designed to become more or less the "standard"
way to pass around coordinates in the library.

This class was written to address a fundamental design issue
that cropped up as I was implementing the LOCPOT reader... namely,
the difference between Direct & Cartesian coordinates, the potential
ambiguity with regards to how coordinates are stored/read, and the
unavoidable fact that each conversion results in lost precision.

Point addresses this problem by tagging coordinate data with the basis
it is expressed in.

(some test cases currently fail due to discrepencies in the range
 of phi and theta; currently deliberating whether this is okay, or if
 Cylindrical/Spherical bases should "canonicalize" their angles)