Tested AnyBasis
It passes the one test I'm probably most concerned about (that
transformations involving AnyBasis produce identical results to
transformations with static bases), so I'm satisfied enough with
it for now to consider getting back to the big project.

Not that the code doesn't require extensive refactoring to make it
even moderately maintainable. But this has side-tracked me for long
enough.